% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandata.R
\name{clean_data}
\alias{clean_data}
\title{Reads data and unselect the specified columns}
\usage{
clean_data(input_path, output_path = NULL, cols = NULL)
}
\arguments{
\item{input_path}{A quoted path to the data file}

\item{output_path}{A quoted destination path}

\item{cols}{A list of quoted column name that we want to unselect}
}
\value{
A data frame excluding the specified columns
}
\description{
Return the data frame without the specified columns from the input_path
}
