% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmax.R
\name{get_max}
\alias{get_max}
\title{Generates a pairplot plotting a scatter chart
and showing correlation between attributes. Also allows a title to be placed}
\usage{
get_max(data, column_name)
}
\arguments{
\item{data}{A dataframe we are going to use}

\item{column_name}{A string representing the column we want to explore}
}
\value{
The max, min and median value
}
\description{
Save the pairplot in the results folder
}
