% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmed.R
\name{get_med}
\alias{get_med}
\title{Generates the minimum value from a column in a dataset}
\usage{
get_med(data, column_name)
}
\arguments{
\item{data}{A dataframe we are going to use}

\item{column_name}{A string representing the column we want to explore}
}
\value{
The min value
}
\description{
Generates the minimum value from a column in a dataset
}
