% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmax.R
\name{get_max}
\alias{get_max}
\title{Generates the maximum value from a column in a dataset}
\usage{
get_max(data, column_name)
}
\arguments{
\item{data}{A dataframe we are going to use}

\item{column_name}{A string representing the column we want to explore}
}
\value{
The max value
}
\description{
Generates the maximum value from a column in a dataset
}
